backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9
# Bot dependencies
ddgs>=6.0
beautifulsoup4>=4.12
requests>=2.31
anthropic>=0.40
jinja2>=3.1
# Testing
pytest>=8.0
httpx>=0.27
